fix extension searchFilters item type
The JSON Schema type for the searchFilters contribution array item should be an object (because it has `name` and `value` properties), but it was incorrectly typed as an array. This type of contribution is rare, which is probably why this bug was not noticed earlier. It would have caused a validation warning when editing the extension's package.json file (in some cases), but otherwise would have had no other negative consequences.